PWN-9581 - fix switching URLs in debug settings
