+++ app/src/main/java/org/p2p/wallet/bridge/repository/BridgeRemoteRepository.kt
@@ -3,21 +3,22 @@ package org.p2p.wallet.bridge.repository
 import com.google.gson.Gson
 import timber.log.Timber
 import java.net.URI
+import org.p2p.core.network.environment.NetworkServicesUrlProvider
 import org.p2p.core.rpc.JsonRpc
 import org.p2p.core.rpc.RpcApi
 import org.p2p.wallet.bridge.api.mapper.BridgeServiceErrorMapper
 import org.p2p.wallet.bridge.model.BridgeResult
-import org.p2p.core.network.environment.NetworkServicesUrlProvider
 
 private const val TAG = "BridgeRemoteRepository"
+
 class BridgeRemoteRepository(
     private val api: RpcApi,
     private val gson: Gson,
     private val errorMapper: BridgeServiceErrorMapper,
-    urlProvider: NetworkServicesUrlProvider,
+    private val urlProvider: NetworkServicesUrlProvider,
 ) : BridgeRepository {
-    private val bridgeStringUrl = urlProvider.loadBridgesServiceEnvironment().baseUrl
-    private val bridgeUrl = URI(bridgeStringUrl)
+    private val bridgeUrl: URI
+        get() = URI(urlProvider.loadBridgesServiceEnvironment().baseUrl)
 
     override suspend fun <P, T> launch(request: JsonRpc<P, T>): BridgeResult.Success<T> {
         try {
