+++ core/src/main/java/org/p2p/core/network/environment/NotificationServiceEnvironment.kt
@@ -1,5 +1,6 @@
 package org.p2p.core.network.environment
 
 data class NotificationServiceEnvironment(
-    val baseUrl: String
+    val baseUrl: String,
+    val isProdSelected: Boolean
 )
