+++ core/src/main/java/org/p2p/core/network/environment/TokenServiceEnvironment.kt
@@ -1,3 +1,6 @@
 package org.p2p.core.network.environment
 
-data class TokenServiceEnvironment(val baseServiceUrl: String)
+data class TokenServiceEnvironment(
+    val baseServiceUrl: String,
+    val isProdSelected: Boolean
+)
